+++ packages/synergy_client_dart/lib/src/interface/interface_export.dart
@@ -1,3 +1,5 @@
+export 'package:synergy_client_dart/src/interface/server_interface.dart';
+export 'package:synergy_client_dart/synergy_client_dart.dart';
 export 'package:synergy_client_dart/src/interface/empty_message_interface.dart';
 export 'package:synergy_client_dart/src/interface/message_interface.dart';
 export 'package:synergy_client_dart/src/interface/screen_interface.dart';