+++ packages/synergy_client_dart/lib/src/synergy_client_dart.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
-import 'dart:io';
 import 'dart:typed_data';
 
+import 'package:synergy_client_dart/src/interface/server_interface.dart';
 import 'package:synergy_client_dart/src/models/server_type.dart';
 import 'package:synergy_client_dart/synergy_client_dart.dart';
 import 'package:synergy_client_dart/src/messages/hello_back_message.dart';
@@ -11,8 +11,10 @@ import 'package:synergy_client_dart/src/messages/message_type.dart';
 import 'package:synergy_client_dart/src/server_proxy.dart';
 import 'package:synergy_client_dart/src/common/extensions.dart';
 
+/// Connect with SynergyServer, we provide by default [SocketServer] but you can provide your own implementation as well
+/// [ServerInterface] is the interface to the screen, to receive all callbacks
 class SynergyClientDart {
-  static Socket? _socket;
+  static ServerInterface? _server;
   static late String _clientName;
   static final ServerProxy _serverProxy = ServerProxy();
   static final _headerMessageSize = MessageHeader.messageTypeSize;
@@ -20,33 +22,33 @@ class SynergyClientDart {
 
   static Future<void> connect({
     required ScreenInterface screen,
-    required String serverIp,
-    required int serverPort,
+    required ServerInterface synergyServer,
     required String clientName,
   }) async {
     _serverProxy.setScreen(screen);
     SynergyClientDart._clientName = clientName;
+    _server = synergyServer;
 
     try {
-      _socket = await Socket.connect(serverIp, serverPort);
-      Logger.log(
-        'Connected to: ${_socket?.remoteAddress.address}:${_socket?.remotePort}',
-      );
-      _serverProxy.screen.onConnect();
-      _socket?.listen(
+      // Dispose previous instance if any
+      await _server?.disconnect();
+
+      // Connect to the new server
+      await _server?.connect(
         _handleSocket,
-        onDone: () {
-          Logger.log('Disconnected from server');
-          _serverProxy.screen.onDisconnect();
-          disconnect();
-        },
-        onError: (error) {
+        (String? error) {
           Logger.log('Error: $error');
           _serverProxy.screen.onError(error.toString());
           disconnect();
         },
+        () {
+          Logger.log('Disconnected from server');
+          _serverProxy.screen.onDisconnect();
+          disconnect();
+        },
       );
-      _serverProxy.dout = _socket;
+      _serverProxy.screen.onConnect();
+      _serverProxy.server = _server;
     } catch (e) {
       Logger.log('Error: $e');
       _serverProxy.screen.onError(e.toString());
@@ -54,9 +56,8 @@ class SynergyClientDart {
   }
 
   static void disconnect() {
-    _socket?.destroy();
-    _socket = null;
-    _serverProxy.dout = null;
+    _server?.disconnect();
+    _serverProxy.server = null;
   }
 
   static void setLogLevel(LogLevel level) {
@@ -70,7 +71,6 @@ class SynergyClientDart {
   static void _handleSocket(Uint8List data) {
     ByteData din = data.buffer.asByteData();
     int messageSize = din.getInt32(0, Endian.big);
-
     // check if the message size is valid
     if (messageSize < 0 || data.length < messageSize) {
       Logger.log(
@@ -87,7 +87,7 @@ class SynergyClientDart {
         Logger.log(helloMessage, LogLevel.debug);
         ServerType serverType = helloMessage.serverType!;
 
-        _socket?.writeMessage(
+        _server?.writeMessage(
           HelloBackMessage(
             majorVersion: 1,
             minorVersion: 6,