+++ packages/synergy_client_dart/lib/synergy_client_dart.dart
@@ -1,5 +1,7 @@
 library synergy_client_dart;
 
+export 'package:synergy_client_dart/src/interface/server_interface.dart';
+export 'package:synergy_client_dart/src/server/socket_server.dart';
 export 'package:synergy_client_dart/src/synergy_client_dart.dart';
 export 'package:synergy_client_dart/src/models/models_export.dart';
 export 'package:synergy_client_dart/src/interface/interface_export.dart';