+++ packages/synergy_client_dart/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.5
+
+- Add SocketServer in connect method
+- Ability to add your own server using ServerInterface
+
 ## 0.0.2
 
 - Update docs